In tools/install.sh code cleanup.

+++ tools/install.sh
@@ -1,8 +1,6 @@
 #!/bin/bash
 
 debugPath=~/Library/Logs/Homebrew
-version=1.0.1
-build=1A007
 
 ls ~/Library/Application\ Support/com.greengecko.brewall 2>/dev/null | grep initializationed > /dev/null 2>&1
 if [ "$?" != "0" ]; then
@@ -40,30 +38,6 @@ if [ $? != 0 ]; then
         echo "Tiger or Leopard 권장"
 
         echo -n "어떤 패키지 관리자를 설치하시겠습니까? (H/t) > "
-        read n
-        if [ "$n" == "t" -o "$n" == "T" ]; then
-            echo "Tigerbrew 설치 방법"
-            echo -e "\033[0;1mhttps://github.com/mistydemeo/tigerbrew\033[m 이 사이트에 들어가서 Tigerbrew를 수동으로 설치하거나"
-            echo "아니면 지금 한번에 설치할 수 있습니다. (제 3자 스크립트를 실행하며 무엇을 할지 설명하고 잠시 대기합니다. )"
-            echo -n "설치하시겠습니까? (Y/n) > "
-            read n
-            if [ "$n" == "n" -o "$n" == "N" ]; then
-                echo "설취를 취소하였습니다. 필수 패키지를 로드할 수 없으므로 종료합니다. "
-                exit 1
-            fi
-            ruby -e "$(curl -fsSkL raw.github.com/mistydemeo/tigerbrew/go/install)"
-        else
-            echo "Homebrew 설치 방법"
-            echo -e "\033[0;1mhttps://brew.sh/index_ko\033[m 이 사이트에 들어가서 Homebrew를 수동으로 설치하거나"
-            echo "아니면 지금 한번에 설치할 수 있습니다. (제 3자 스크립트를 실행하며 무엇을 할지 설명하고 잠시 대기합니다. )"
-            echo -n "설치하시겠습니까? (Y/n) > "
-            read n
-            if [ "$n" == "n" -o "$n" == "N" ]; then
-                echo "설취를 취소하였습니다. 필수 패키지를 로드할 수 없으므로 종료합니다. "
-                exit 1
-            fi
-            /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)"
-        fi
     else
         echo "This brewall script require brew. Because extend of these tools."
         
@@ -77,30 +51,52 @@ if [ $? != 0 ]; then
         echo "Tiger or Leopard recommand"
 
         echo -n "What would you like install brew package? (H/t) > "
-        read n
-        if [ "$n" == "t" -o "$n" == "T" ]; then
+    fi
+    read n
+    if [ "$n" == "t" -o "$n" == "T" ]; then
+        if [ $LANG == "ko_KR.UTF-8" ]; then
+            echo "Tigerbrew 설치 방법"
+            echo -e "\033[0;1mhttps://github.com/mistydemeo/tigerbrew\033[m 이 사이트에 들어가서 Tigerbrew를 수동으로 설치하거나"
+            echo "아니면 지금 한번에 설치할 수 있습니다. (제 3자 스크립트를 실행하며 무엇을 할지 설명하고 잠시 대기합니다. )"
+            echo -n "설치하시겠습니까? (Y/n) > "
+        else
             echo "Tigerbrew Installation guide"
             echo -e "Please enter this site \033[0;1mhttps://github.com/mistydemeo/tigerbrew\033[m and manual install Tigerbrew or "
             echo "Install now on this script. (Execute Third party script and explains what it will do and then pauses before it does it. )"
             echo -n "Install Tigerbrew now? (Y/n) > "
-            read n
-            if [ "$n" == "n" -o "$n" == "N" ]; then
+        fi
+        read n
+        if [ "$n" == "n" -o "$n" == "N" ]; then
+            if [ $LANG == "ko_KR.UTF-8" ]; then
+                echo "설취를 취소하였습니다. 필수 패키지를 로드할 수 없으므로 종료합니다. "
+            else
                 echo "Installation aborted. Can not load require package, terminating."
-                exit 1
             fi
-            ruby -e "$(curl -fsSkL raw.github.com/mistydemeo/tigerbrew/go/install)"
+            exit 1
+        fi
+        ruby -e "$(curl -fsSkL raw.github.com/mistydemeo/tigerbrew/go/install)"
+    else
+        if [ $LANG == "ko_KR.UTF-8" ]; then
+            echo "Homebrew 설치 방법"
+            echo -e "\033[0;1mhttps://brew.sh/index_ko\033[m 이 사이트에 들어가서 Homebrew를 수동으로 설치하거나"
+            echo "아니면 지금 한번에 설치할 수 있습니다. (제 3자 스크립트를 실행하며 무엇을 할지 설명하고 잠시 대기합니다. )"
+            echo -n "설치하시겠습니까? (Y/n) > "
         else
             echo "Homebrew Installation guide"
             echo -e "Please enter this site \033[0;1mhttps://brew.sh\033[m and manual install Homebrew or "
             echo "Install now on this script. (Execute Third party script and explains what it will do and then pauses before it does it. )"
             echo -n "Install Homebrew now? (Y/n) > "
-            read n
-            if [ "$n" == "n" -o "$n" == "N" ]; then
-                echo "Installation aborted. Can not load require package, terminating."
-                exit 1
-            fi
-            /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)"
         fi
+        read n
+        if [ "$n" == "n" -o "$n" == "N" ]; then
+            if [ $LANG == "ko_KR.UTF-8" ]; then
+                echo "설취를 취소하였습니다. 필수 패키지를 로드할 수 없으므로 종료합니다. "
+            else
+                echo "Installation aborted. Can not load require package, terminating."
+            fi
+            exit 1
+        fi
+        /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)"
     fi
 fi
